restructured popups
