+++ lorem.txt
@@ -3,3 +3,5 @@
 Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis non elit volutpat, faucibus lacus quis, rhoncus odio. Phasellus lacinia enim ac imperdiet scelerisque. Nunc ac viverra elit. Sed tristique pulvinar scelerisque. Integer sed rhoncus enim, vel facilisis mauris. Vivamus quis commodo neque. Morbi feugiat ante vel neque finibus, euismod commodo augue euismod.
 djdjdjdjdt
 iererrotuerotiuerdfn,kgdkjghdkfnckbjvonkjbnkj
+
+
