+++ src/lib.rs
@@ -1,13 +1,12 @@
 //expose modules
 pub mod app;
 pub mod editor;
+pub mod popups;
 pub mod cursor;
 pub mod config;
-pub mod popup;
 pub mod command_line;
 pub mod clipboard;
 pub mod errors;
 pub mod ui;
 pub mod input;
-pub mod error_popup;
-pub mod confirmation_popup;
+
