+++ src/main.rs
@@ -13,13 +13,12 @@ pub mod app; //expose app module
 mod clipboard;
 mod command_line;
 mod config;
-mod confirmation_popup;
 mod cursor;
 // Declare the 'editor' module. This makes the editor/ directory (via mod.rs) visible.
 mod editor;
-mod error_popup;
+mod popups;
 pub mod input; //expose input module
-mod popup;
+
 
 pub mod ui;
 mod errors;
