+++ src/popups/mod.rs
@@ -1,0 +1,4 @@
+
+pub mod popup;
+pub mod error_popup;
+pub mod confirmation_popup;
